Library

Mitosis.AffineMap — Type

AffineMap(B, β)

Represents a function f = AffineMap(B, β) such that f(x) == B*x + β.

Mitosis.BF — Type

BF()

Backward filter for linear Gaussian systems parametrized by mean and covariance of the backward filtered marginal distribution.

Mitosis.BFFG — Type

BFFG()

Backward filter forward guiding context for non-linear Gaussian systems with h parametrized by WGaussian{(:F,:Γ,:c)}` (see Theorem 7.1 [Automatic BFFG].)

Mitosis.ConstantMap — Type

ConstantMap(β)

Represents a function f = ConstantMap(β) such that f(x) == β.

Mitosis.Gaussian — Type

Gaussian{(:μ,:Σ)}
Gaussian{(:F,:Γ)}

Mitosis provides the measure Gaussian based on MeasureTheory.jl, with a mean μ and covariance Σ parametrization, or parametrised by natural parameters F = Γ μ, Γ = Σ⁻¹.

Usage:

Gaussian(μ=m, Σ=C)
p = Gaussian{(:μ,:Σ)}(m, C)
Gaussian(F=C\m, Γ=inv(C))

convert(Gaussian{(:F,:Γ)}, p)

rand(rng, p)

Mitosis.LinearMap — Type

LinearMap(B)

Represents a function f = LinearMap(B) such that f(x) == B*x.

MeasureTheory.kernel — Function

kernel(f, M)
kernel((f1, f2, ...), M)

A kernel κ = kernel(f, M) returns a wrapper around a function f giving the parameters for a measure of type M, such that κ(x) = M(f(x)...) respective κ(x) = M(f1(x), f2(x), ...).

If the argument is a named tuple (;a=f1, b=f1), κ(x) is defined as M(;a=f(x),b=g(x)).

Reference

https://en.wikipedia.org/wiki/Markov_kernel

Mitosis.conditional — Method

conditional(p::Gaussian, A, B, xB)

Conditional distribution of X[i for i in A] given X[i for i in B] == xB if $X ~ P$.

Mitosis.correct — Method

correct(prior, obskernel, obs) = u, yres, S

Joseph form correction step of a Kalman filter with prior state and obs the observation with observation kernel obskernel = kernel(Gaussian; μ=LinearMap(H), Σ=ConstantMap(R)) H is the observation operator and R the observation covariance. Returns corrected/conditional distribution u, the residual and the innovation covariance. See https://en.wikipedia.org/wiki/Kalman_filter#Update.